fix: update seeder routes to fix validate gotchas
